1. Added Browser Factory 2. Removed thread sleeps
