+++ src/main/java/yale/pageObjects/BasePage.java
@@ -2,16 +2,15 @@ package yale.pageObjects;
 
 import framework.BaseElement;
 import framework.Browser;
-import framework.Config;
 import org.openqa.selenium.By;
 
-public class BasePage  {
+public class BasePage {
 
-    protected Browser browser =Browser.getInstance();
+    protected Browser browser = Browser.getInstance();
     protected BaseElement baseElement;
 
     public BasePage(By by) {
-        baseElement= new BaseElement();
+        baseElement = new BaseElement();
         BaseElement.getElement(by);
     }
 }
