+++ src/main/java/yale/pageObjects/MainPage.java
@@ -25,6 +25,9 @@ public class MainPage extends BasePage {
     private final By FIND_PEOPLE_LINK = By.xpath("//span[contains(text(), 'Find People')]");
     private final By A_Z_FACULTY_LINK = By.xpath("//span[contains(text(), 'A-Z Faculty List')]");
     private final By ANCHOR_ALPHABET_LINK = By.xpath("//ul[@class='category-anchors-list category-anchors-list--alphabet']//button[@tabindex=0]");
+    private final By SIDEBAR_NAVIGATION_GIVING = By.xpath("//span[@class='sidebar-navigation__link-text'][contains(text(), 'Giving')]");
+    private final By SELECT_EVENT_BUTTON = By.xpath("//span[contains(text(), 'Select')]");
+    private final By CONTACT_US_TITLE = By.xpath("//h1[@class='generic-content__title']");
 
     public MainPage() {
         super(YSM_LOGO);
@@ -32,25 +35,21 @@ public class MainPage extends BasePage {
 
     public MainPage openMainPage() {
         Log.logInfo("Open Main Page");
-        //browser.getUrl(config.getProperty("mainPageURL"));
         return this;
     }
 
     public void clickPatientCareLink() {
         Log.logInfo("Click Patient Care Link");
-        BaseElement.waitForElementToBePresent(LINK_TO_PATIENT_CARE_PAGE);
         baseElement.clickElement(LINK_TO_PATIENT_CARE_PAGE);
     }
 
     public boolean searchPanelIsDisplayed() {
         Log.logInfo("Check that Search Panel is displayed");
-        BaseElement.waitForElementToBePresent(SEARCH_PANEL);
         return baseElement.isElementDisplayed(SEARCH_PANEL);
     }
 
     public boolean ysmLogoIsDisplayed() {
         Log.logInfo("Check that YSM logo is displayed");
-        BaseElement.waitForElementToBePresent(YSM_LOGO);
         return baseElement.isElementDisplayed(YSM_LOGO);
     }
 
@@ -61,13 +60,11 @@ public class MainPage extends BasePage {
 
     public void clickSiteIcon() {
         Log.logInfo("Click Site Icon");
-        BaseElement.waitForElementToBePresent(YSM_LOGO);
         baseElement.clickElement(YSM_LOGO);
     }
 
     public void hoverOnNavigationLink() {
         Log.logInfo("Hover on navigation Link");
-        BaseElement.waitForElementToBePresent(NAVIGATION_LINK);
         baseElement.hoverOnElement(NAVIGATION_LINK);
     }
 
@@ -79,34 +76,32 @@ public class MainPage extends BasePage {
 
     public void clickSupportUsLink() {
         Log.logInfo("Click Support Us Link");
-        BaseElement.waitForElementToBePresent(SUPPORT_US_LINK);
         baseElement.clickElement(SUPPORT_US_LINK);
+        BaseElement.waitForElementToBeClickable(SIDEBAR_NAVIGATION_GIVING);
     }
 
     public MainPage clickCalendarLink() {
         Log.logInfo("Click Calendar Link");
-        BaseElement.waitForElementToBePresent(CALENDAR_LINK);
         baseElement.clickElement(CALENDAR_LINK);
+        BaseElement.waitForElementToBeClickable(SELECT_EVENT_BUTTON);
         return this;
     }
 
     public MainPage clickLogo() {
         Log.logInfo("Click Logo");
         baseElement.clickElement(YSM_LOGO);
-        BaseElement.waitForElementToBePresent(YSM_LOGO);
         return this;
     }
 
     public MainPage clickContactUsLink() {
         Log.logInfo("Click Contact Us Link");
-        BaseElement.waitForElementToBePresent(CONTACT_US_LINK);
         baseElement.clickElement(CONTACT_US_LINK);
+        BaseElement.waitForText(CONTACT_US_TITLE, "Contact YSM");
         return this;
     }
 
     public NewsPage clickNewsTitle() {
         Log.logInfo("Click News Title");
-        BaseElement.waitForElementToBePresent(NEWS_TITLE);
         baseElement.clickElement(NEWS_TITLE);
         return new NewsPage();
     }
@@ -118,14 +113,12 @@ public class MainPage extends BasePage {
 
     public SearchPage clickMoreTopStoriesButton() {
         Log.logInfo("Click More Top Stories Button");
-        BaseElement.waitForElementToBeClickable(MORE_TOP_STORIES);
         baseElement.clickElement(MORE_TOP_STORIES);
         return new SearchPage();
     }
 
     public SearchPage clickMoreEventsButton() {
         Log.logInfo("Click More Events Button");
-        BaseElement.waitForElementToBeClickable(SEE_ALL_EVENTS_BUTTON);
         baseElement.clickElement(SEE_ALL_EVENTS_BUTTON);
         return new SearchPage();
     }
@@ -139,8 +132,7 @@ public class MainPage extends BasePage {
     public String getTitleWebsitePage() {
         Log.logInfo("Get Web Site Title");
         String title = baseElement.getTitle();
-        String websiteTitle = title.substring(title.lastIndexOf("<") + 1).trim();
-        return websiteTitle;
+        return title.substring(title.lastIndexOf("<") + 1).trim();
     }
 
     public boolean isClinicalTrialHeaderIsDisplayed() {
@@ -151,54 +143,49 @@ public class MainPage extends BasePage {
 
     public EventPage openFirstEvent() {
         Log.logInfo("Open first event");
-        BaseElement.waitForTheFirstElementFromArrayIsClickable(EVENT_TITLE);
         baseElement.clickFirstElementFromArray(EVENT_TITLE);
         return new EventPage();
     }
 
     public String getFirstEventTitle() {
+        Log.logInfo("Get first event info");
         BaseElement.waitForTheFirstElementFromArrayIsClickable(EVENT_TITLE);
-        String eventTitle = baseElement.getTextFirstElementFromArray(EVENT_TITLE);
-        return eventTitle;
+        return baseElement.getTextFirstElementFromArray(EVENT_TITLE);
     }
 
     public MainPage goToTwitter() {
         Log.logInfo("Go to twitter");
-        BaseElement.waitForElementToBeClickable(TWITTER_FEED);
         baseElement.clickElement(TWITTER_FEED);
         return this;
     }
 
     public MainPage goAToAboutYSMPage() {
         Log.logInfo("Go to AboutYSMPage");
-        BaseElement.waitForElementToBeClickable(ABOUT_YSM_LINK);
         baseElement.clickWithJS(ABOUT_YSM_LINK);
         return this;
     }
 
     public MainPage goToFindPeoplePage() {
         Log.logInfo("Go to Find People Page");
-        BaseElement.waitForElementToBeClickable(FIND_PEOPLE_LINK);
         baseElement.clickWithJS(FIND_PEOPLE_LINK);
         return this;
     }
 
     public MainPage goToA_ZFacultyPage() {
         Log.logInfo("Go to A_Z Faculty Page");
-        BaseElement.waitForElementToBeClickable(A_Z_FACULTY_LINK);
         baseElement.clickWithJS(A_Z_FACULTY_LINK);
         return this;
     }
 
     public String clickAnyAnchorLink() {
         Log.logInfo("Click Any Anchor Link");
-        BaseElement.waitForTheFirstElementFromArrayIsClickable(ANCHOR_ALPHABET_LINK);
-        String anchorLetter = baseElement.chooseValueFromArrayAndGetText(ANCHOR_ALPHABET_LINK);
-        return anchorLetter;
+        return baseElement.chooseValueFromArrayAndGetText(ANCHOR_ALPHABET_LINK);
     }
 
     public boolean isTheFirstLetterCoincidesWithTheChosenAnchorLink(String chosenLetter) {
         Log.logInfo("Get Text Active Element");
         return baseElement.isChosenValueCoincidesWithTheFirstLetterOnFocus(chosenLetter);
     }
+
+
 }
