+++ src/main/java/yale/pageObjects/SearchOverlay.java
@@ -1,14 +1,14 @@
 package yale.pageObjects;
 
 import framework.BaseElement;
+import framework.Browser;
+import framework.logger.Log;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebElement;
-import framework.logger.Log;
-import framework.Browser;
 
 import java.util.List;
 
-public class SearchOverlay extends MainPage {
+public class SearchOverlay extends SearchPage {
 
     private final By SEARCH_INPUT = By.xpath("//input[@type='search']");
     private final By SEARCH_BUTTON = By.xpath("//button[@aria-label='Go to Search Page']");
@@ -23,7 +23,7 @@ public class SearchOverlay extends MainPage {
     private final By FIRST_SEARCH_RESULT = By.xpath("//ul[@class='search-results-list__search-results']//a[@class='search-result']");
     private final By PAGE_TITLE = By.xpath("//h1[@class='profile-details-header__name']");
     private final By PEOPLE_NAME_INPUT = By.xpath("//input[@id='peopleName']");
-    private final By SEARCH_COUNT = By.xpath("//h3[@class='find-people-tab__results-count']");
+    private final static By SEARCH_COUNT = By.xpath("//h3[@class='find-people-tab__results-count']");
     private final By RESEARCH_AREA_INPUT = By.xpath("//input[@id='researchArea']");
     private final By RESEARCH_AREA_OPTION = By.xpath("//ul[@class='research-area-input__suggestions-container']//li[@role='option']");
     private final By ROLE_DROP_DOWN = By.xpath("//div[@class='find-people-tab__roles-dropdown']");
@@ -31,29 +31,17 @@ public class SearchOverlay extends MainPage {
 
     public SearchPage clickSearchButton() {
         Log.logInfo("Click Search button");
-        BaseElement.waitForElementToBePresent(SEARCH_BUTTON);
         baseElement.clickElement(SEARCH_BUTTON);
-        BaseElement.waitForElementToBePresent(SEARCH_INPUT);
+        BaseElement.waitForElementToBeClickable(PEOPLE_FILTER);
         return new SearchPage();
     }
 
-    public int getSearchCount() throws InterruptedException {
+    public int getSearchCount() {
         Log.logInfo("Get search count");
-        BaseElement.waitForElementToBePresent(SEARCH_COUNT);
-        Thread.sleep(3000);
         String searchCount = baseElement.getText(SEARCH_COUNT);
         String[] stringArray = searchCount.split(" ");
         String stringNumber = stringArray[0];
-        int number = Integer.parseInt(stringNumber);
-        System.out.println(number);
-        return number;
-    }
-
-    public MainPage inputSearchValue(String value) {
-        Log.logInfo("Input search value" + value);
-        BaseElement.waitForElementToBePresent(SEARCH_INPUT);
-        baseElement.typeTo(SEARCH_INPUT, value);
-        return this;
+        return Integer.parseInt(stringNumber);
     }
 
     public boolean isInputSearchPanelDisplayed() {
@@ -64,25 +52,21 @@ public class SearchOverlay extends MainPage {
 
     public boolean isFindPeopleTabDisplayed() {
         Log.logInfo("Check that Find People Tab is displayed");
-        BaseElement.waitForElementToBePresent(FIND_PEOPLE_TAB);
         return baseElement.isElementDisplayed(FIND_PEOPLE_TAB);
     }
 
     public boolean isPopularLinksTabDisplayed() {
         Log.logInfo("Check that Popular Links Tab is displayed");
-        BaseElement.waitForElementToBePresent(POPULAR_LINKS_TAB);
         return baseElement.isElementDisplayed(POPULAR_LINKS_TAB);
     }
 
     public boolean isSearchTabDisplayed() {
         Log.logInfo("Check that Search Tab is displayed");
-        BaseElement.waitForElementToBePresent(SEARCH_TAB);
         return baseElement.isElementDisplayed(SEARCH_TAB);
     }
 
     public boolean isSearchResultDetailsDisplayed() {
-        Log.logInfo("Check that search result details is displayed");
-        BaseElement.waitForElementsFromArrayArePresent(SEARCH_RESULT_DETAILS);
+        Log.logInfo("Check that Search Result Details is displayed");
         return baseElement.isElementsDisplayed(SEARCH_RESULT_DETAILS);
     }
 
@@ -109,37 +93,23 @@ public class SearchOverlay extends MainPage {
         return baseElement.compareText(SEARCH_RESULT_TITLE, HIGHLIGHTED_TEXT, searchValue);
     }
 
-    public SearchOverlay openFirstResult() {
+    public ProfilePage openFirstPeopleResult() {
         Log.logInfo("Open First Result");
-        List<WebElement> elements = Browser.getDriver().findElements(FIRST_SEARCH_RESULT);
-        WebElement element = elements.get(0);
-        BaseElement.waitForTheFirstElementFromArrayIsClickable(FIRST_SEARCH_RESULT);
         baseElement.clickFirstElementFromArray(FIRST_SEARCH_RESULT);
-        return this;
-    }
-
-    public boolean isTitleContainsSearchValue(String searchValue) {
-        Log.logInfo("Compare title with inputted search value " + searchValue);
-        BaseElement.waitForElementToBePresent(PAGE_TITLE);
-        String title = baseElement.getTitle();
-        System.out.println(title);
-        if (title.contains(searchValue)) {
-            return true;
-        }
-        return false;
+        return new ProfilePage();
     }
 
     public SearchOverlay clickFindPeopleTab() {
         Log.logInfo("Click Find People Tab");
-        BaseElement.waitForElementToBePresent(FIND_PEOPLE_TAB);
         baseElement.clickElement(FIND_PEOPLE_TAB);
         return this;
     }
 
     public SearchOverlay typePeopleName(String value) {
         Log.logInfo("Type people Name " + value);
-        BaseElement.waitForElementToBePresent(PEOPLE_NAME_INPUT);
+        String firstText = SearchOverlay.getSearchResultCount();
         baseElement.typeTo(PEOPLE_NAME_INPUT, value);
+        BaseElement.waitForInvisibility(SEARCH_COUNT, firstText);
         return this;
     }
 
@@ -151,8 +121,9 @@ public class SearchOverlay extends MainPage {
 
     public SearchOverlay clickSuggestionForResearchArea() {
         Log.logInfo("Click suggestion for research area");
-        BaseElement.waitForTheFirstElementFromArrayIsClickable(RESEARCH_AREA_OPTION);
+        String firstText = SearchOverlay.getSearchResultCount();
         baseElement.clickFirstElementFromArray(RESEARCH_AREA_OPTION);
+        BaseElement.waitForInvisibility(SEARCH_COUNT, firstText);
         return this;
     }
 
@@ -164,7 +135,13 @@ public class SearchOverlay extends MainPage {
 
     public void addRoleOption() {
         Log.logInfo("Add Role Option");
-        BaseElement.waitForTheFirstElementFromArrayIsClickable(ROLE_OPTION);
+        String firstText = SearchOverlay.getSearchResultCount();
         baseElement.clickTheSecondElementFromArray(ROLE_OPTION);
+        BaseElement.waitForInvisibility(SEARCH_COUNT, firstText);
+    }
+
+    public static String getSearchResultCount() {
+        Log.logInfo("Get Search Result Count");
+        return BaseElement.getText(SEARCH_COUNT);
     }
 }
