+++ src/main/java/yale/services/OpenSearchPage.java
@@ -9,5 +9,6 @@ public class OpenSearchPage {
         mainPage.openMainPage();
         mainPage.clickPerformSearch()
                 .clickSearchButton();
+
     }
 }
