+++ src/main/resources/yale/config.properties
@@ -1,2 +1,3 @@
 driverPath=src/main/resources/yale/chromedriver 3
 mainPageURL=https://acceptance.medicine.yale.edu/
+browser=chrome
